+++ app/src/main/java/com/kiero/presentation/auth/parent/component/KakaoLoginButton.kt
@@ -1,4 +1,4 @@
-package com.kiero.presentation.auth.component
+package com.kiero.presentation.auth.parent.component
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
@@ -20,6 +20,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.kiero.R
 import com.kiero.core.designsystem.theme.KieroTheme
 
 
@@ -44,7 +45,7 @@ fun KakaoLoginButton(
         elevation = ButtonDefaults.buttonElevation(defaultElevation = 0.dp)
     ) {
         Icon(
-            painter = painterResource(id = com.kiero.R.drawable.ic_kakao_login),
+            painter = painterResource(id = R.drawable.ic_kakao_login),
             contentDescription = null,
             modifier = Modifier.size(24.dp),
             tint = KieroTheme.colors.black
